requirements.txt
django
djangorestframework
requests


https://github.com/mateusschwede/django/tree/main/CursoDjangoAPI